Postpone window event adapter assertion
The previous implementation asserted that the adapter had been set when
the events where being registered in GLFW.

This caused to crash because the adapter is registered after the GLFW
event handlers are registered.

This only temporary fix for the issue, if an event is assigned during
registration but before adapter assignment this will fail.